Refactored components for better readability. #13
